Update as_sparse6() matrix method, docs and tests
Add as_sparse6.matrix(n=) to expect network size (number of vertices),
which references #22. The methods for igraphs and networks set this
argument when calling the matrix method. By necessity the generic and
the methods gained `...` too.

The tests have been updated accordingly.

On the way updated and extended the documentation.